Set the package version in heat-dist.conf [revision]
Heat has an API call to allow users to discover the version of heat
being consumed. It can be called with the command:

  heat build-info

However the implementation simply returns the conf value
[revision] heat_revision, so it is up to downstream installers
to set this value to something useful.

This change sets the value in /usr/share/heat/heat-dist.conf to
%{version}.

As someone who supports heat having this information available can
save a lot of time in diagnosing already fixed issues.

Change-Id: Iea3e43fea9553f49f4b388bad21e501f0299856b