Seperate entrypoint for tempest plugins
Previously openstack-heat-commons package installed the entry point for
the in-tree Tempest plugin which cause failure while running tempest list-plugins
because Tempest found the entry point but does not have Tempest plugin
code which is subpackaged in python-heat-tests.
Separate heat_tests egg-info is created with the plugin entry point
so that Tempest includes the plugin only when subpackage is installed.

Change-Id: I0dc2d2d7bd98df208fa906fcf6dc95c13a747dbc