Subpackage python-heat-tests-tempest
This creates the subpackage python-heat-tests-tempest which packages
heat_integrationtests and registers an entry point for a tempest test
plugin.

This differs from the other subpackages as the tempest entry point
can't be registered in the openstack-heat-common package since this
will break tempest when openstack-heat-common is installed but
python-heat-tests-tempest isn't.

Adding to the difficulty of this problem is that setuptools and pbr
have a strong assumption to one package per repository.

The solution proposed here is to do a standard python install then
duplicate the egg-info directory and manipulate it to resemble a
dedicated tempest plugin package.

Change-Id: I0dc2d2d7bd98df208fa906fcf6dc95c13a747dbc