Add a separate "monolith" subpackage for heat-all
The new heat-all binary is generally not the recommended way of running
Heat, so we probably don't want it in the openstack-heat-engine
subpackage.

This change adds a separate openstack-heat-monolith subpackage
containing just the heat-all binary, and adds a systemd service file for
it. However, this subpackage is *not* required by the main
openstack-heat package, so users who need it can request it
specifically and others will not have it installed.

Change-Id: Ibba4c90aeee8c51d8d233387fe2555d4d000ebed