Don't set server URLs to localhost in heat-dist.conf
The correct URLs for the metadata server &c. can now be inferred from
the Keystone catalog, as long as nothing is specified in the config
files. Prior to this it made sense to default the values to localhost,
but now it hurts more than it helps.

Change-Id: Id8977e04e1e522fc2c79c7f509dfa8bea61ecd05
(cherry picked from commit b0b9bff492de21b26dd14dd2c0b128a8b61ae9a9)