Update patch set 4
Patch Set 4: Code-Review+1

Usually we do both:
Obsoletes: %{name}-api-cloudwatch < %{epoch}:10.0.0
Provides: %{name}-api-cloudwatch = %{epoch}:%{version}-%{release}

That usually helps the dep solver but makes more sense when we actually provides some kind of replacement. So it's up to you whether to remove or keep the provides.

Also that would prevent the 'obsolete-not-provided' warning from rpmlint:
"If a package is obsoleted by a compatible replacement, the obsoleted package should also be provided in order to not cause unnecessary dependency breakage.
If the obsoleting package is not a compatible replacement for the old one, leave out the Provides."

Patch-set: 4
Reviewer: Gerrit User 169 <169@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1