Store Heat api-paste.ini in a distro agnostic dir
Historically as of Liberty release of RDO, we keep storing
api-paste-dist.ini in the distro specific path under /usr/share/.
That is done as a downstream-only change in Oslo config.

This does not work well as we likely loose that downstream-only
patch for each new release and we also cannot contribute that distro
specific path for upstream Heat and Oslo config.

Switch it to be distributed as /etc/heat/heat.conf.d/api-paste.ini,
which is supported for Oslo's config_dir.

Change-Id: Ia5b24f9cd54ad9550606119d4d115c7e52928946
Related: rhbz#1796098
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>