rhosp: Make clients for unsupported components optional
The following components are not supported in RHOSP.
 - Magnum
 - Mistral
 - Sahara
 - Trove

This change removes client package for these unsupported services from
dependent packages. Note that Zaqar is no longer supported, but it's
still kept because Heat uses zaqar in signaling and the feature would
fail if zaqarclient is missing.

Also, dependencies on the client packages are removed from the -test
package because the required packages are pulled by the common package
which the -test package requires.

Change-Id: If2ae135fb0eaaf12d890115a6cbcd0ac385c2b8f